ml/synthetic_invoice_generator/requirements.txt
faker>=20.0.0
reportlab>=4.0.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.22.0
imgaug>=0.4.0
babel>=2.12.0
jinja2>=3.1.0
pyyaml>=6.0
PyMuPDF>=1.23.0

